do not import prisma in edge bundle for virtual tickets
